Moved searchTestSuite next to its only usage
